# openapi.yaml
openapi: 3.0.3
info:
  title: Second-Hand Shop Scraper API
  description: |
    REST API for searching and retrieving products from Czech second-hand shops.
    
    This API provides access to search history and product results from multiple
    Czech second-hand marketplaces including:
    - Bazos.cz
    - Sbazar.cz
    - Avizo.cz
    - Inzeruj.cz
    - Aukro.cz
    - eBay (Browse API)

  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com

servers:
  - url: http://localhost:8091/api/v1
    description: Local development server
  - url: http://localhost:8091/api/v1
    description: Docker container

tags:
  - name: Health
    description: Health check endpoints
  - name: Searches
    description: Search history and management
  - name: Products
    description: Product listings from searches

paths:
  /health:
    get:
      tags:
        - Health
      summary: Health check
      description: Check if the API is running and responsive
      operationId: healthCheck
      responses:
        '200':
          description: API is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  time:
                    type: string
                    format: date-time
                    example: "2026-02-03T10:30:00Z"

  /searches:
    get:
      tags:
        - Searches
      summary: Get all searches
      description: Retrieve a list of all search queries that have been performed
      operationId: getSearches
      responses:
        '200':
          description: List of searches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Search'
              example:
                - id: 1
                  keyword: "hemingway"
                  created_at: "2026-02-03T10:00:00Z"
                  updated_at: "2026-02-03T10:00:00Z"
                - id: 2
                  keyword: "rypadlo"
                  created_at: "2026-02-03T11:00:00Z"
                  updated_at: "2026-02-03T11:00:00Z"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - Searches
      summary: Track a new keyword
      description: |
        Saves a new keyword to track. If the keyword is already tracked,
        returns the existing search rather than erroring (creation is
        idempotent). Responds as soon as the search is saved - it does not
        wait for results. Matching listings are fetched in the background
        right away, and again on every scheduled cron run afterwards.
      operationId: createSearch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSearchRequest'
            example:
              keyword: "joy-con pair"
      responses:
        '201':
          description: Search saved (or already existed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Search'
        '400':
          description: Missing or invalid keyword
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /searches/{searchId}:
    delete:
      tags:
        - Searches
      summary: Stop tracking a search
      description: |
        Deletes a saved search. Previously scraped products are not
        deleted - only this search's association with them.
      operationId: deleteSearch
      parameters:
        - name: searchId
          in: path
          required: true
          description: ID of the search
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        '204':
          description: Search deleted
        '400':
          description: Invalid search ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Search not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /searches/{searchId}/products:
    get:
      tags:
        - Products
      summary: Get products for a search
      description: Retrieve all products found for a specific search query
      operationId: getSearchProducts
      parameters:
        - name: searchId
          in: path
          required: true
          description: ID of the search
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        '200':
          description: Search with products
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchWithProducts'
        '400':
          description: Invalid search ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Search not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    Search:
      type: object
      required:
        - id
        - keyword
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier for the search
          example: 1
        keyword:
          type: string
          description: The search keyword used
          example: "hemingway"
        created_at:
          type: string
          format: date-time
          description: Timestamp when the search was created
          example: "2026-02-03T10:00:00Z"
        updated_at:
          type: string
          format: date-time
          description: Timestamp when the search was last updated
          example: "2026-02-03T10:00:00Z"

    CreateSearchRequest:
      type: object
      required:
        - keyword
      properties:
        keyword:
          type: string
          description: The keyword to track, 1-255 characters
          example: "joy-con pair"

    Product:
      type: object
      required:
        - id
        - title
        - price
        - currency
        - url
        - shop_source
        - auction_type
        - condition
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier for the product
          example: 1
        title:
          type: string
          description: Product title
          example: "Hemingwayové - John Hemingway"
        description:
          type: string
          description: Product description
          example: "Krásná kniha v dobrém stavu"
        price:
          type: number
          format: float
          description: Product price
          example: 30.0
        currency:
          type: string
          description: Currency code
          example: "CZK"
        url:
          type: string
          format: uri
          description: Product URL
          example: "https://www.bazos.cz/inzerat/214456837/hemingway-kniha.php"
        image_url:
          type: string
          format: uri
          description: Product image URL
          example: "https://www.bazos.cz/images/123456.jpg"
        location:
          type: string
          description: Product location
          example: "Praha"
        shop_source:
          type: string
          description: Source marketplace
          enum:
            - bazos.cz
            - sbazar.cz
            - avizo.cz
            - inzeruj.cz
            - aukro.cz
            - ebay.com
          example: "bazos.cz"
        auction_type:
          type: string
          description: Type of listing
          enum:
            - auction
            - sale
          example: "sale"
        condition:
          type: string
          description: Product condition
          enum:
            - new
            - used
            - like_new
            - good
            - fair
            - poor
            - damaged
            - unknown
          example: "used"
        ending_time:
          type: string
          format: date-time
          nullable: true
          description: Auction ending time (for auctions only)
          example: "2026-02-05T18:00:00Z"
        created_at:
          type: string
          format: date-time
          description: Timestamp when the product was first scraped
          example: "2026-02-03T10:00:00Z"
        updated_at:
          type: string
          format: date-time
          description: Timestamp when the product was last updated
          example: "2026-02-03T10:00:00Z"

    SearchWithProducts:
      type: object
      required:
        - search
        - products
        - total
      properties:
        search:
          $ref: '#/components/schemas/Search'
        products:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        total:
          type: integer
          description: Total number of products
          example: 29
      example:
        search:
          id: 1
          keyword: "hemingway"
          created_at: "2026-02-03T10:00:00Z"
          updated_at: "2026-02-03T10:00:00Z"
        products:
          - id: 1
            title: "Hemingwayové - John Hemingway"
            description: "Kniha"
            price: 30.0
            currency: "CZK"
            url: "https://www.bazos.cz/inzerat/214456837/hemingway.php"
            image_url: ""
            location: "Praha"
            shop_source: "bazos.cz"
            auction_type: "sale"
            condition: "used"
            ending_time: null
            created_at: "2026-02-03T10:00:00Z"
            updated_at: "2026-02-03T10:00:00Z"
        total: 1

    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
          example: "Search not found"
        message:
          type: string
          description: Detailed error message
          example: "No search found with ID: 999"